Set window masks during window creation
Instead of making a separate call to set window masks, set them during
creation. Also replace the erroneous ButtonPress with ButtonPressMask.
The former is meant as a event type while the letter is an event mask.
They happen to have the same value, but this is not true for most events